LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-eval-reporting-report-only-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: The Content Security Policy 'require-trusted-types-for 'script'' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: Refused to load because it does not appear in the object-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

TIMEOUT Trusted Type violation report: evaluating a string. Test timed out
NOTRUN Trusted Type violation report: evaluating a Trusted Script.
NOTRUN Trusted Type violation report: default policy runs in report-only mode.
